refactor(kimi-k3): isolate model ops and trim PR scope
Move Kimi-specific kernels and metadata into model_ops, remove inactive debug and generic changes, and keep the implementation focused on the validated TP8 path.
