+++ atom/examples/simple_inference.py
@@ -59,9 +59,7 @@ def main():
     engine_args = EngineArgs.from_cli_args(args)
     llm = engine_args.create_engine()
 
-    tokenizer = AutoTokenizer.from_pretrained(
-        args.model, trust_remote_code=args.trust_remote_code
-    )
+    tokenizer = AutoTokenizer.from_pretrained(args.model)
 
     sampling_params = SamplingParams(
         temperature=args.temperature, max_tokens=args.max_tokens
