+++ atom/model_ops/sampler.py
@@ -9,8 +9,6 @@ from aiter.ops.triton.softmax import softmax
 from aiter.ops.triton.topk import topk
 from torch import nn
 
-from atom.utils import envs
-
 # Try to import aiter top-k/top-p sampling ops
 try:
     import aiter.ops.sampling  # noqa: F401
@@ -75,11 +73,6 @@ class Sampler(nn.Module):
         Returns:
             Sampled token IDs (num_tokens,)
         """
-        if envs.ATOM_DEBUG_TOPK > 0:
-            from atom.utils.debug_helper import maybe_log_topk
-
-            maybe_log_topk(logits, prefix="sampler ")
-
         # No Top-K Top-P parameters, perform temperature-based sampling
         if not self._needs_filtering(top_ks, top_ps):
             return self._temperature_sample(
